+++ functions/src/helpers/helpers.ts
