+++ functions/src/helpers/setup.ts
