+++ functions/src/triggers/database.ts
@@ -1,0 +1,1 @@
+// TODO: Add database triggers (e.g. adding admin permissions)
